feat(app): removed update social media and refactored file names
