+++ src/sections/social/social-table-head.tsx
@@ -1,6 +1,5 @@
 import Box from '@mui/material/Box';
 import TableRow from '@mui/material/TableRow';
-import Checkbox from '@mui/material/Checkbox';
 import TableHead from '@mui/material/TableHead';
 import TableCell from '@mui/material/TableCell';
 import TableSortLabel from '@mui/material/TableSortLabel';
@@ -9,7 +8,7 @@ import { visuallyHidden } from './utils';
 
 // ----------------------------------------------------------------------
 
-type UserTableHeadProps = {
+type SocialTableHeadProps = {
   orderBy: string;
   rowCount: number;
   numSelected: number;
@@ -27,7 +26,7 @@ export function UserTableHead({
   headLabel,
   numSelected,
   onSelectAllRows,
-}: UserTableHeadProps) {
+}: SocialTableHeadProps) {
   return (
     <TableHead>
       <TableRow>
