+++ src/sections/social/utils.ts
@@ -1,4 +1,4 @@
-import type { UserProps } from './user-table-row';
+import type { SocialProps } from './social-table-row';
 
 // ----------------------------------------------------------------------
